move pages to pages folder
