+++ project/src/pages/favorites-screen/favorites-screen.tsx
