+++ project/src/pages/main-screen/main-screen.tsx
@@ -1,4 +1,4 @@
-import PlaceCard from '../place-card/place-card';
+import PlaceCard from '../../components/place-card/place-card';
 
 type MainScreenProps = {
   cardsCount: number;
